+++ CHANGELOG.md
@@ -20,4 +20,10 @@
 
 - Rename `Result<T>.ToType` to `To` and allow attempting to convert to non-derived types.
 
-- Add ToString extensions supporting `IFormattable`.
+- Add `ToString` extensions supporting `IFormattable`.
+
+- Add `ResultEqualityComparer<T>`.
+
+- Add additional overloads for `Result<T>.Equals` which allow specifying an equality comparer to use for comparing values.
+
+- Rename `GetValueOrDefault(T @default)` and `GetValueOrDefault(Func<T> getDefault)` to `GetValueOr`.